src/agents/ladbs/pyproject.toml
[project]
name = "ladbs-agent"
version = "0.1.0"
description = "LADBS AI Agent for Azure AI Foundry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core",
    "agent-framework-azure-ai",
    "azure-identity",
    "azure-ai-projects",
    "pyyaml>=6.0",
]

[tool.uv]
package = false
